Replaced FloatField with DecimalField for financial amounts.

+++ trunk/plugins/priority/models.py
@@ -34,7 +34,7 @@ class UserPriority(models.Model):
     expire = models.DateTimeField()
     txn_id = models.CharField(max_length=40)
     currency = models.CharField(max_length=3)
-    payment = models.FloatField()
+    payment = models.DecimalField(max_digits=7, decimal_places=2)
 
     class Admin:
         list_display = ('user', 'priority', 'activated', 'expire',
